+++ index.html
@@ -18,7 +18,7 @@
     <div id="calendar-next"></div>
   </div>
 </body>
-<script src="./scripts/app.js"></script>
+<script src="./build/app.js"></script>
 </body>
 
 </html>
