+++ package.json
@@ -1,0 +1,29 @@
+{
+  "name": "month",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "build": "webpack"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ozhu91/calendar.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/ozhu91/calendar/issues"
+  },
+  "homepage": "https://github.com/ozhu91/calendar#readme",
+  "dependencies": {
+    "webpack": "^4.41.5",
+    "@babel/cli": "^7.8.3",
+    "@babel/core": "^7.8.3",
+    "@babel/preset-env": "^7.8.3",
+    "babel-loader": "^8.0.6"
+  },
+  "devDependencies": {
+    "webpack-cli": "^3.3.10"
+  }
+}
